backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
chromadb>=1.5.0
numpy>=2.2
openai==1.59.6
pymupdf==1.25.3
python-dotenv==1.0.1
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
